Moved pdf search from doc view mode to pdf viewer
