+++ ui/src/components/Document/DocumentViewMode.jsx
@@ -1,10 +1,8 @@
 import React, { lazy, Suspense } from 'react';
-import { defineMessages, injectIntl } from 'react-intl';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router';
 
-import Query from 'app/Query';
 import AudioViewer from 'viewers/AudioViewer';
 import DefaultViewer from 'viewers/DefaultViewer';
 import TableViewer from 'viewers/TableViewer';
@@ -14,7 +12,6 @@ import ImageViewer from 'viewers/ImageViewer';
 import FolderViewer from 'viewers/FolderViewer';
 import EmailViewer from 'viewers/EmailViewer';
 import VideoViewer from 'viewers/VideoViewer';
-import EntityActionBar from 'components/Entity/EntityActionBar';
 import { SectionLoading } from 'components/common';
 import { selectEntityDirectionality } from 'selectors';
 
@@ -22,48 +19,9 @@ import './DocumentViewMode.scss';
 
 const PdfViewer = lazy(() => import(/* webpackChunkName: 'base' */ 'src/viewers/PdfViewer'));
 
-const messages = defineMessages({
-  placeholder: {
-    id: 'entity.viewer.search_placeholder',
-    defaultMessage: 'Search in {label}',
-  },
-});
-
 export class DocumentViewMode extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.onSearch = this.onSearch.bind(this);
-  }
-
-  onSearch(queryText) {
-    const { history, location, query } = this.props;
-    const newQuery = query.setString('q', queryText);
-
-    history.push({
-      pathname: location.pathname,
-      search: newQuery.toLocation(),
-    });
-  }
-
-  renderActionBar() {
-    const { document, query, intl } = this.props;
-
-    if (document.schema.isA('Pages')) {
-      return (
-        <EntityActionBar
-          query={query}
-          onSearchSubmit={this.onSearch}
-          searchPlaceholder={intl.formatMessage(messages.placeholder, { label: document.getCaption() })}
-          searchDisabled={document.getProperty('processingError')?.length}
-        />
-      );
-    }
-    return null;
-  }
-
   renderContent() {
-    const { document, queryText, activeMode, dir } = this.props;
+    const { document, activeMode, dir } = this.props;
     const processingError = document.getProperty('processingError');
 
     if (processingError && processingError.length) {
@@ -126,7 +84,6 @@ export class DocumentViewMode extends React.Component {
         <Suspense fallback={<SectionLoading />}>
           <PdfViewer
             document={document}
-            queryText={queryText}
             activeMode={activeMode}
             dir={dir}
           />
@@ -145,7 +102,6 @@ export class DocumentViewMode extends React.Component {
   render() {
     return (
       <div className="DocumentViewMode">
-        {this.renderActionBar()}
         {this.renderContent()}
       </div>
     );
@@ -154,17 +110,13 @@ export class DocumentViewMode extends React.Component {
 
 
 const mapStateToProps = (state, ownProps) => {
-  const { location, document } = ownProps;
-  const query = Query.fromLocation('entities', location, {}, '');
+  const { document } = ownProps;
   return {
     dir: selectEntityDirectionality(state, document),
-    query,
-    queryText: query.getString('q'),
   };
 };
 
 export default compose(
   withRouter,
   connect(mapStateToProps),
-  injectIntl,
 )(DocumentViewMode);
